requirements.txt
pytest
pytest-playwright
playwright
pandas
openpyxl
python-dotenv